# frontend/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Momentum Terminal API",
    "description": "US equities 12-1 momentum research and INTERNAL paper simulation. No brokerage connectivity.",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/ledger/orders": {
      "get": {
        "tags": [
          "ledger"
        ],
        "summary": "Orders",
        "operationId": "orders_api_ledger_orders_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OrdersPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ledger/fills": {
      "get": {
        "tags": [
          "ledger"
        ],
        "summary": "Fills",
        "operationId": "fills_api_ledger_fills_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FillsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ledger/cash": {
      "get": {
        "tags": [
          "ledger"
        ],
        "summary": "Cash",
        "operationId": "cash_api_ledger_cash_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CashPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/ledger/events": {
      "get": {
        "tags": [
          "ledger"
        ],
        "summary": "Events",
        "operationId": "events_api_ledger_events_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 2000,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "level",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Level"
            }
          },
          {
            "name": "category",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventsPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CashPage": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/CashTx"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "rows"
        ],
        "title": "CashPage"
      },
      "CashTx": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "session": {
            "type": "string",
            "title": "Session"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "symbol": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Symbol"
          },
          "amount": {
            "type": "number",
            "title": "Amount"
          },
          "balance_after": {
            "type": "number",
            "title": "Balance After"
          },
          "fill_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fill Id"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session",
          "kind",
          "symbol",
          "amount",
          "balance_after",
          "fill_id",
          "note",
          "created_at"
        ],
        "title": "CashTx"
      },
...
      "EventModel": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "ts": {
            "type": "string",
            "title": "Ts"
          },
          "level": {
            "type": "string",
            "enum": [
              "info",
              "warning",
              "error"
            ],
            "title": "Level"
          },
          "category": {
            "type": "string",
            "title": "Category"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "portfolio_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Portfolio Id"
          },
          "run_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Run Id"
          },
          "session": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Session"
          },
          "payload": {
            "additionalProperties": true,
            "type": "object",
            "title": "Payload"
          }
        },
        "type": "object",
        "required": [
          "id",
          "ts",
          "level",
          "category",
          "message",
          "portfolio_id",
          "run_id",
          "session",
          "payload"
        ],
        "title": "EventModel"
      },
      "EventsPage": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/EventModel"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "rows"
        ],
        "title": "EventsPage"
      },
      "FillModel": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "session": {
            "type": "string",
            "title": "Session"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "side": {
            "type": "string",
            "enum": [
              "buy",
              "sell"
            ],
            "title": "Side"
          },
          "shares": {
            "type": "integer",
            "title": "Shares"
          },
          "ref_price": {
            "type": "number",
            "title": "Ref Price"
          },
          "fill_price": {
            "type": "number",
            "title": "Fill Price"
          },
          "gross_value": {
            "type": "number",
            "title": "Gross Value"
          },
          "slippage_cost": {
            "type": "number",
            "title": "Slippage Cost"
          },
          "commission": {
            "type": "number",
            "title": "Commission"
          },
          "plan_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Id"
          },
          "order_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order Id"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "session",
          "symbol",
          "side",
          "shares",
          "ref_price",
          "fill_price",
          "gross_value",
          "slippage_cost",
          "commission"
        ],
        "title": "FillModel"
      },
      "FillsPage": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/FillModel"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "rows"
        ],
        "title": "FillsPage"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "OrderModel": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "plan_id": {
            "type": "integer",
            "title": "Plan Id"
          },
          "symbol": {
            "type": "string",
            "title": "Symbol"
          },
          "intended_side": {
            "type": "string",
            "title": "Intended Side"
          },
          "target_weight": {
            "type": "number",
            "title": "Target Weight"
          },
          "est_shares": {
            "type": "integer",
            "title": "Est Shares"
          },
          "fill_session": {
            "type": "string",
            "title": "Fill Session"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "status_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Status Reason"
          },
          "filled_shares": {
            "type": "integer",
            "title": "Filled Shares"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "plan_id",
          "symbol",
          "intended_side",
          "target_weight",
          "est_shares",
          "fill_session",
          "status",
          "status_reason",
          "filled_shares",
          "created_at",
          "updated_at"
        ],
        "title": "OrderModel"
      },
      "OrdersPage": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          },
          "rows": {
            "items": {
              "$ref": "#/components/schemas/OrderModel"
            },
            "type": "array",
            "title": "Rows"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limit",
          "offset",
          "rows"
        ],
        "title": "OrdersPage"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}